moved uv inversion to base output handler and added option to do uv inversion for netcdf writes. This adds an invert_uv argument to ForwardPassStrategy
